term with coeff base class
