+++ lib/node/PolynomialTerm.js
@@ -1,3 +1,4 @@
+const TermWithCoefficient = require('./TermWithCoefficient');
 const NodeCreator = require('./Creator');
 const NodeType = require('./Type');
 
@@ -14,18 +15,20 @@ const evaluate = require('../util/evaluate');
  - exponent: a node that can take any form, e.g. x^(2+x^2)
    (might be null if no exponent)
 */
-class PolynomialTerm {
+class PolynomialTerm extends TermWithCoefficient {
   // if onlyImplicitMultiplication is true, an error will be thrown if `node`
   // is a polynomial term without implicit multiplication
   // (i.e. 2*x instead of 2x) and therefore isPolynomialTerm will return false.
   constructor(node, onlyImplicitMultiplication=false) {
+    super(node, onlyImplicitMultiplication);
+
     if (NodeType.isOperator(node)) {
       if (node.op === '^') {
         const symbolNode = node.args[0];
         if (!NodeType.isSymbol(symbolNode)) {
           throw Error('Expected symbol term, got ' + symbolNode);
         }
-        this.symbol = symbolNode;
+        this.base = symbolNode;
         this.exponent = node.args[1];
       }
       // it's '*' ie it has a coefficient
@@ -48,7 +51,7 @@ class PolynomialTerm {
           throw Error('Cannot have two coefficients ' + coeffNode +
             ' and ' + nonCoefficientTerm.getCoeffNode());
         }
-        this.symbol = nonCoefficientTerm.getSymbolNode();
+        this.base = nonCoefficientTerm.getSymbolNode();
         this.exponent = nonCoefficientTerm.getExponentNode();
       }
       // this means there's a fraction coefficient
@@ -64,7 +67,7 @@ class PolynomialTerm {
           throw Error('Polynomial terms cannot have nested fractions');
         }
         this.exponent = numeratorNode.getExponentNode();
-        this.symbol = numeratorNode.getSymbolNode();
+        this.base = numeratorNode.getSymbolNode();
         const numeratorConstantNode = numeratorNode.getCoeffNode(true);
         this.coeff = NodeCreator.operator(
           '/', [numeratorConstantNode, denominatorNode]);
@@ -81,7 +84,7 @@ class PolynomialTerm {
       const polyNode = new PolynomialTerm(
         arg, onlyImplicitMultiplication);
       this.exponent = polyNode.getExponentNode();
-      this.symbol = polyNode.getSymbolNode();
+      this.base = polyNode.getSymbolNode();
       if (!polyNode.hasCoeff()) {
         this.coeff = NodeCreator.constant(-1);
       }
@@ -90,7 +93,7 @@ class PolynomialTerm {
       }
     }
     else if (NodeType.isSymbol(node)) {
-      this.symbol = node;
+      this.base = node;
     }
     else {
       throw Error('Unsupported node type: ' + node.type);
@@ -99,58 +102,11 @@ class PolynomialTerm {
 
   /* GETTER FUNCTIONS */
   getSymbolNode() {
-    return this.symbol;
+    return this.base;
   }
 
   getSymbolName() {
-    return this.symbol.name;
-  }
-
-  getCoeffNode(defaultOne=false) {
-    if (!this.coeff && defaultOne) {
-      return NodeCreator.constant(1);
-    }
-    else {
-      return this.coeff;
-    }
-  }
-
-  getCoeffValue() {
-    if (this.coeff) {
-      return evaluate(this.coeff);
-    }
-    else {
-      return 1; // no coefficient is like a coeff of 1
-    }
-  }
-
-  getExponentNode(defaultOne=false) {
-    if (!this.exponent && defaultOne) {
-      return NodeCreator.constant(1);
-    }
-    else {
-      return this.exponent;
-    }
-  }
-
-  getRootNode() {
-    return NodeCreator.polynomialTerm(
-      this.symbol, this.exponent, this.coeff);
-  }
-
-  // note: there is no exponent value getter function because the exponent
-  // can be any expression and not necessarily a number.
-
-  /* CHECKER FUNCTIONS (returns true / false for certain conditions) */
-
-  // Returns true if the coefficient is a fraction
-  hasFractionCoeff() {
-    // coeffNode is either a constant or a division operation.
-    return this.coeff && NodeType.isOperator(this.coeff);
-  }
-
-  hasCoeff() {
-    return !!this.coeff;
+    return this.base.name;
   }
 }
 
@@ -169,18 +125,5 @@ PolynomialTerm.isPolynomialTerm = function(
     return false;
   }
 };
-
-// Multiplies `node`, a constant or fraction of two constant nodes, by -1
-// Returns a node
-function negativeCoefficient(node) {
-  if (NodeType.isConstant(node)) {
-    node = NodeCreator.constant(0 - parseFloat(node.value));
-  }
-  else {
-    const numeratorValue = 0 - parseFloat(node.args[0].value);
-    node.args[0] = NodeCreator.constant(numeratorValue);
-  }
-  return node;
-}
 
 module.exports = PolynomialTerm;
