+++ lib/util/print.js
@@ -27,7 +27,7 @@ function printTreeTraversal(node, parentNode) {
       const coeffStr = printTreeTraversal(coeffTerm);
 
       const nonCoeffTerm = Node.Creator.polynomialTerm(
-        polyTerm.symbol, polyTerm.exponent, null);
+        polyTerm.getSymbolNode(), polyTerm.exponent, null);
       const nonCoeffStr = printTreeTraversal(nonCoeffTerm);
 
       return `${coeffStr} ${nonCoeffStr}`;
